move docs utils
