+++ app/lib/docs-utils/github.server.ts
@@ -1,4 +1,4 @@
-import { env } from "@/utils/env.server";
+import { env } from "@/lib/docs-utils/env.server";
 import { Octokit } from "octokit";
 
 export const octokit = new Octokit(env.GITHUB_TOKEN ? { auth: env.GITHUB_TOKEN } : undefined);
