+++ app/routes/_landing.docs.$lang.$ref.$.tsx
@@ -4,10 +4,10 @@ import { siteConfig } from "@/config/site";
 import iconsHref from "@/icons.svg";
 import type { loader as rootLoader } from "@/root";
 import { type loader as docsLayoutLoader, useGitHubRef } from "@/routes/_landing.docs.$lang.$ref";
-import type { Doc } from "@/utils/github";
-import { getRepoDoc } from "@/utils/github";
-import { CACHE_CONTROL, handleRedirects } from "@/utils/http.server";
-import { getMeta } from "@/utils/meta";
+import type { Doc } from "@/lib/docs-utils/github";
+import { getRepoDoc } from "@/lib/docs-utils/github";
+import { CACHE_CONTROL, handleRedirects } from "@/lib/docs-utils/http.server";
+import { getMeta } from "@/lib/docs-utils/meta";
 import type { HeadersFunction, LoaderFunctionArgs, SerializeFrom } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
